Renamed and merged all *UnitTests into *Tests.
