--- core/src/test/java/org/hisp/dhis/android/core/configuration/ConfigurationModelUnitTests.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) 2017, University of Oslo
- *
- * All rights reserved.
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
- package org.hisp.dhis.android.core.configuration;
-
-import org.junit.Test;
-
-import nl.jqno.equalsverifier.EqualsVerifier;
-import nl.jqno.equalsverifier.Warning;
-
-import static junit.framework.Assert.fail;
-
-public class ConfigurationModelUnitTests {
-
-    @Test
-    public void equals_shouldConformToContract() {
-        ConfigurationModel configurationModel = ConfigurationModel.builder()
-                .serverUrl("").build();
-
-        // ToDo: report issue in EqualsVerifier library, which fails the test even
-        // ToDo: though NULL_FIELDS warning is suppressed
-//        EqualsVerifier.forClass(configurationModel.getClass())
-//                .suppress(Warning.NULL_FIELDS)
-//                .verify();
-    }
-
-    @Test
-    public void build_shouldFailOnNullServerUrl() {
-        try {
-            ConfigurationModel.builder().build();
-
-            fail("IllegalStateException was expected but nothing was thrown");
-        } catch (IllegalStateException illegalStateException) {
-            // swallow exception
-        }
-    }
-}
