+++ core/src/test/java/org/hisp/dhis/android/core/dataelement/DataElementTests.java
@@ -38,9 +38,19 @@ import org.junit.Test;
 import java.io.IOException;
 import java.text.ParseException;
 
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
+
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
 public class DataElementTests {
+
+    @Test
+    public void equals_shouldConformToContract() {
+        EqualsVerifier.forClass(DataElementModel.builder().build().getClass())
+                .suppress(Warning.NULL_FIELDS)
+                .verify();
+    }
 
     @Test
     public void dataElement_shouldMapFromJsonString() throws IOException, ParseException {
