+++ core/src/test/java/org/hisp/dhis/android/core/option/OptionSetTests.java
@@ -38,9 +38,19 @@ import org.junit.Test;
 import java.io.IOException;
 import java.text.ParseException;
 
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
+
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
 public class OptionSetTests {
+
+    @Test
+    public void equals_shouldConformToContract() {
+        EqualsVerifier.forClass(OptionSetModel.builder().build().getClass())
+                .suppress(Warning.NULL_FIELDS)
+                .verify();
+    }
 
     @Test
     public void optionSet_shouldMapFromJsonString() throws IOException, ParseException {
