+++ core/src/test/java/org/hisp/dhis/android/core/user/AuthenticatedUserModelTests.java
@@ -36,7 +36,7 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 
 @RunWith(JUnit4.class)
-public class AuthenticatedUserModelUnitTests {
+public class AuthenticatedUserModelTests {
 
     @Test
     public void equals_shouldConformToContract() {
