+++ core/src/test/java/org/hisp/dhis/android/core/user/UserCredentialsModelTests.java
@@ -26,17 +26,22 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
- package org.hisp.dhis.android.core.program;
+ package org.hisp.dhis.android.core.user;
 
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 
-public class ProgramStageSectionUnitTests {
+// ToDo: consider testing .isValid() as well.
+@RunWith(JUnit4.class)
+public class UserCredentialsModelTests {
+
     @Test
     public void equals_shouldConformToContract() {
-        EqualsVerifier.forClass(ProgramStageSectionModel.builder().build().getClass())
+        EqualsVerifier.forClass(UserCredentialsModel.builder().build().getClass())
                 .suppress(Warning.NULL_FIELDS)
                 .verify();
     }
