+++ core/src/test/java/org/hisp/dhis/android/core/user/UserOrganisationUnitLinkModelTests.java
@@ -35,13 +35,12 @@ import org.junit.runners.JUnit4;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 
-// ToDo: consider testing .isValid() as well.
 @RunWith(JUnit4.class)
-public class UserModelUnitTests {
+public class UserOrganisationUnitLinkModelTests {
 
     @Test
     public void equals_shouldConformToContract() {
-        EqualsVerifier.forClass(UserModel.builder().build().getClass())
+        EqualsVerifier.forClass(UserOrganisationUnitLinkModel.builder().build().getClass())
                 .suppress(Warning.NULL_FIELDS)
                 .verify();
     }
